chore(builder/citrea): Move to different folders.
