+++ core/src/builder/citrea/request.rs
@@ -9,12 +9,10 @@
 use crate::errors::BridgeError;
 use bitcoin::consensus::Encodable;
 use bitcoin::hashes::Hash;
-use bitcoin::{Block, Transaction, Txid};
+use bitcoin::{Transaction, Txid};
 use jsonrpsee::core::client::ClientT;
 use jsonrpsee::http_client::HttpClient;
 use jsonrpsee::rpc_params;
-use merkle::{Hashable, MerkleTree, Proof};
-use ring::digest::SHA256;
 use serde_json::json;
 
 const CITREA_ADDRESS: &str = "0x3100000000000000000000000000000000000002";
@@ -54,37 +52,6 @@ macro_rules! encode_btc_params {
             .flatten()
             .collect::<Vec<u8>>()
     };
-}
-
-fn create_merkle_proof<T>(data: Vec<T>, index: usize) -> Result<Proof<T>, BridgeError>
-where
-    T: Hashable + Clone,
-{
-    let merkle_tree = MerkleTree::from_vec(&SHA256, data.clone());
-
-    let proof = merkle_tree.gen_nth_proof(index);
-
-    proof.ok_or(BridgeError::Error("TODO".to_string()))
-}
-
-pub async fn prepare_deposit_params(block: Block, target_txid: Txid) -> Result<(), BridgeError> {
-    let mut txid_index = 0;
-    let txids = block
-        .txdata
-        .iter()
-        .enumerate()
-        .map(|(i, tx)| {
-            if tx.compute_txid() == target_txid {
-                txid_index = i;
-            }
-
-            tx.compute_txid() // or is this wtxid?
-        })
-        .collect::<Vec<_>>();
-
-    let _merkle_proof = create_merkle_proof(txids, txid_index)?;
-
-    Ok(())
 }
 
 /// # Parameters
@@ -141,7 +108,6 @@ pub async fn deposit(
     Ok(())
 }
 
-/// TODO: withdrawal_index is u256, convert it to [u8; 32]
 pub async fn withdrawal_utxos(
     client: HttpClient,
     withdrawal_index: u32,
